change server

+++ server-side.c
@@ -20,9 +20,9 @@ float delay = 0;
 // This function is responsible for, given a packet size, computing the delay (for links with speed=1200 bit/s)
 // WARNING: the packet size is given in bytes so a conversion to bits is needed
 float link_delay_tx(int packet_size) {
-	printf("Received packet size: %i\n", packet_size);
-	double delay = (packet_size * 8)/LINK_SPEED;
-	printf("Generated delay: %f for packet size %lf\n", delay, packet_size);
+	printf("Received packet size: %d\n", packet_size);
+	float delay = (packet_size * 8)/LINK_SPEED;
+	printf("Generated delay: %f for packet size %d\n", delay, packet_size);
 	return delay;
 }
